+++ src/app/modules/comparisons/comparisons.page.ts
@@ -2,19 +2,21 @@ import { Component, OnInit } from '@angular/core';
 import { ComparisonService } from 'src/app/services/comparison.service';
 import { IModel } from 'src/app/interfaces/model.interface';
 import { KeyValue } from '@angular/common';
+import { AbstractGoToTopPage } from 'src/app/utils/abstract-go-to-top-page';
 
 @Component({
   selector: 'app-comparisons',
   templateUrl: './comparisons.page.html',
   styleUrls: ['./comparisons.page.scss'],
 })
-export class ComparisonsPage implements OnInit {
+export class ComparisonsPage extends AbstractGoToTopPage implements OnInit {
 
   public comparisons: IModel[];
 
   constructor(
     private comparisonsService: ComparisonService
   ) {
+    super();
     this.comparisons = [];
   }