+++ src/app/modules/favorites/favorites.page.ts
@@ -1,12 +1,15 @@
 import { Component } from '@angular/core';
+import { AbstractGoToTopPage } from 'src/app/utils/abstract-go-to-top-page';
 
 @Component({
   selector: 'app-favorites',
   templateUrl: 'favorites.page.html',
   styleUrls: ['favorites.page.scss'],
 })
-export class FavoritesPage {
+export class FavoritesPage extends AbstractGoToTopPage {
 
-  constructor() { }
+  constructor() {
+    super();
+  }
 
 }