+++ src/app/modules/searches/searches.page.ts
@@ -1,12 +1,15 @@
 import { Component } from '@angular/core';
+import { AbstractGoToTopPage } from 'src/app/utils/abstract-go-to-top-page';
 
 @Component({
   selector: 'app-searches',
   templateUrl: 'searches.page.html',
   styleUrls: ['searches.page.scss'],
 })
-export class SearchesPage {
+export class SearchesPage extends AbstractGoToTopPage {
 
-  constructor() { }
+  constructor() {
+    super();
+  }
 
 }