+++ src/app/utils/abstract-go-to-top-page.ts
@@ -1,0 +1,10 @@
+import { ViewChild } from '@angular/core';
+import { IonContent } from '@ionic/angular';
+
+export class AbstractGoToTopPage {
+    @ViewChild(IonContent, { static: false }) content: IonContent;
+
+    public goToTop(): void {
+        this.content.scrollToTop(1000);
+    }
+}