+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 - Added `ModelPage`
 - Added some info to `ModelPage`
+- Added scroll to top functionality
 
 ## [0.4.6proto]
 ### Changed